Update pitonstream.
This adds more parameters to override the default board tile config (e.g. a dual core Genesys 2 mapping) when running bare-metal tests over pitonstream.

Further, this adds an ifdef around the loads to the magic PASS/FAIL addresses for pitonstream. This is needed, since this might otherwise to a blocked system when debugging via the RISC-V DTM + JTAG.