Change uart reg-shift in device tree to 0
Xilinx UART 16550 has 4 byte register boundaries which is not directly NS16550 compatible. We already modified the hardware to shift out the extra bits to 1 byte boundaries. This reg-shift field wasn't read by Linux but will be read by OpenSBI so this patch fixes it to the value it should be (which is 0).